:bug: Fix DELETE/qrcode endpoint

+++ app/dao/entities/qrcode.py
@@ -1,3 +1,5 @@
+import copy
+
 from flask import jsonify
 from sqlalchemy import Column, String, Integer, ForeignKey, DateTime
 import secrets
@@ -16,7 +18,7 @@ def qrcodes_get_method(session):
         return {'error': "QrCode not found.",
                 'message': "Aucun Qrcode disponible."}, 404
     else:
-        return jsonify([x.serialize for x in qrcodes]), 200
+        return jsonify([x.serialize(True) for x in qrcodes]), 200
 
 
 def qrcode_get_method(session, id):
@@ -25,7 +27,7 @@ def qrcode_get_method(session, id):
         return {'error': "QrCode not found.",
                 'message': "Ce QrCode n'existe pas."}, 404
 
-    return jsonify(qrcode.serialize), 200
+    return jsonify(qrcode.serialize(True)), 200
 
 
 def qrcode_get_method_by_token(session, token):
@@ -34,7 +36,7 @@ def qrcode_get_method_by_token(session, token):
         return {'error': "Unknown token",
                 'message': "Token inconnu."}, 400
     else:
-        return jsonify(signatures[0].serialize), 200
+        return jsonify(signatures[0].serialize(True)), 200
 
 
 def qrcode_post_method(json, session):
@@ -64,7 +66,8 @@ def qrcode_post_method(json, session):
 
     session.add(qrcode)
     session.commit()
-    return jsonify(qrcode.serialize), 201
+
+    return jsonify(qrcode.serialize(True)), 201
 
 
 def qrcode_patch_method(json, session, id):
@@ -91,13 +94,20 @@ def qrcode_patch_method(json, session, id):
 
 def qrcode_delete_method(session, id):
     qrcode = session.query(Qrcode).get(id)
+
+    if not qrcode:
+        return {'error': "QrCode not found.",
+                'message': ""}, 404
+
     signatures = session.query(signature.Signature).filter(signature.Signature.token == qrcode.token).all()
     for sign in signatures:
         session.delete(sign)
+        session.commit()
 
     session.delete(qrcode)
+    session.commit()
 
-    return jsonify(qrcode.serialize)
+    return jsonify(qrcode.serialize(False)), 200
 
 
 class Qrcode(Entity, Base):
@@ -106,7 +116,7 @@ class Qrcode(Entity, Base):
     title = Column(String)
     token = Column(String, unique=True, nullable=False)
     created_by_id = Column(Integer, ForeignKey('catho_user.id'))
-    created_by = relationship("CathoUser")
+    created_by = relationship("CathoUser", lazy='subquery')
     start_time = Column(DateTime)
     end_time = Column(DateTime)
 
@@ -118,13 +128,12 @@ class Qrcode(Entity, Base):
         self.start_time = start_time
         self.end_time = end_time
 
-    @property
-    def serialize(self):
+    def serialize(self, serialize_user):
         return {
             'id': self.id,
             'title': self.title,
             'token': self.token,
-            'user': self.created_by.serialize,
+            'user': self.created_by.serialize if serialize_user else self.created_by_id,
             'start_time': self.start_time,
             'end_time': self.end_time
         }
